+++ src/App.tsx
@@ -12,9 +12,9 @@ import SignupPage from "./pages/SignupPage";
 import NoticePage from "./pages/NoticePage";
 import NoticeDetailPage from "./pages/NoticeDetailPage";
 import PointLayout from "./layouts/PointLayout";
-import MainList from "./components/point/MainList";
-import PointCharge from "./components/point/PointCharge";
-import PointDetail from "./components/point/PointDetail";
+import MainList from "./components/myPage/point/MainList";
+import PointCharge from "./components/myPage/point/PointCharge";
+import PointDetail from "./components/myPage/point/PointDetail";
 import HotSpaceListPage from "./pages/HotSpaceListPage";  
 import ProfilePage from "./pages/ProfilePage";
 import WithdrawalPage from "./pages/WithdrawalPage";
