+++ src/components/myPage/notice/ContentSection.tsx
