+++ src/components/myPage/point/DetailItem.tsx
