+++ src/components/myPage/point/DetailSection.tsx
