+++ src/components/myPage/point/MyPointSection.tsx
