+++ src/components/myPage/point/PointDetail.tsx
